trial a new BaseData object
