+++ specparam/objs/base.py
@@ -6,18 +6,18 @@ import numpy as np
 from numpy.linalg import LinAlgError
 from scipy.optimize import curve_fit
 
+from specparam.objs.data import BaseData
 from specparam.core.utils import group_three
 from specparam.core.strings import gen_width_warning_str
 from specparam.core.funcs import gaussian_function, get_ap_func
-from specparam.core.errors import DataError, InconsistentDataError, NoDataError, FitError
-from specparam.utils.data import trim_spectrum
+from specparam.core.errors import NoDataError, FitError
 from specparam.utils.params import compute_gauss_std
 from specparam.sim.gen import gen_aperiodic, gen_periodic
 
 ###################################################################################################
 ###################################################################################################
 
-class BaseSpectralModel():
+class BaseSpectralModel(BaseData):
     """Base object defining model & algorithm for spectral parameterization.
 
     Parameters
@@ -52,12 +52,6 @@ class BaseSpectralModel():
         Run mode: whether the object is set in debug mode.
         If in debug mode, an error is raised if model fitting is unsuccessful.
         This should be controlled by using the `set_debug_mode` method.
-    _check_freqs : bool
-        Run mode: whether to check the frequency values.
-        If True, checks the frequency values, and raises an error for uneven spacing.
-    _check_data : bool
-        Run mode: whether to check the power spectrum values.
-        If True, checks the power values and raises an error for any NaN / Inf values.
 
     Attributes
     ----------
@@ -80,9 +74,10 @@ class BaseSpectralModel():
                  ap_percentile_thresh=0.025, ap_guess=(None, 0, None),
                  ap_bounds=((-np.inf, -np.inf, -np.inf), (np.inf, np.inf, np.inf)),
                  cf_bound=1.5, bw_std_edge=1.0, gauss_overlap_thresh=0.75,
-                 maxfev=5000, error_metric='MAE',
-                 debug_mode=False, check_freqs_mode=True, check_data_mode=True):
+                 maxfev=5000, error_metric='MAE', debug_mode=False):
         """Initialize base model object"""
+
+        BaseData.__init__(self)
 
         ## Public settings
         self.peak_width_limits = peak_width_limits
@@ -104,19 +99,10 @@ class BaseSpectralModel():
 
         ## RUN MODES
         self._debug = debug_mode
-        self._check_freqs = check_freqs_mode
-        self._check_data = check_data_mode
 
         ## Set internal settings, based on inputs, and initialize data & results attributes
         self._reset_internal_settings()
         self._reset_data_results(True, True, True)
-
-
-    @property
-    def has_data(self):
-        """Indicator for if the object contains data."""
-
-        return True if np.any(self.power_spectrum) else False
 
 
     @property
@@ -133,38 +119,25 @@ class BaseSpectralModel():
 
         return True if not np.all(np.isnan(self.aperiodic_params_)) else False
 
-
     def add_data(self, freqs, power_spectrum, freq_range=None, clear_results=True):
         """Add data (frequencies, and power spectrum values) to the current object.
 
         Parameters
         ----------
-        freqs : 1d array
-            Frequency values for the power spectrum, in linear space.
-        power_spectrum : 1d array
-            Power spectrum values, which must be input in linear space.
-        freq_range : list of [float, float], optional
-            Frequency range to restrict power spectrum to.
-            If not provided, keeps the entire range.
+        % copied in from Data object
         clear_results : bool, optional, default: True
             Whether to clear prior results, if any are present in the object.
             This should only be set to False if data for the current results are being re-added.
 
         Notes
         -----
-        If called on an object with existing data and/or results
-        they will be cleared by this method call.
+        % copied in from Data object
         """
 
-        # If any data is already present, then clear previous data
-        # Also clear results, if present, unless indicated not to
-        #   This is to ensure object consistency of all data & results
-        self._reset_data_results(clear_freqs=self.has_data,
-                                 clear_spectrum=self.has_data,
-                                 clear_results=self.has_model and clear_results)
+        # Clear results, if present, unless indicated not to
+        self._reset_results(clear_results=self.has_model and clear_results)
 
-        self.freqs, self.power_spectrum, self.freq_range, self.freq_res = \
-            self._prepare_data(freqs, power_spectrum, freq_range, 1)
+        super().add_data(freqs, power_spectrum, freq_range=None)
 
 
     def fit(self, freqs=None, power_spectrum=None, freq_range=None):
@@ -261,7 +234,7 @@ class BaseSpectralModel():
 
             # Clear any interim model results that may have run
             #   Partial model results shouldn't be interpreted in light of overall failure
-            self._reset_data_results(clear_results=True)
+            self._reset_results(clear_results=True)
 
             # Print out status
             if self.verbose:
@@ -289,26 +262,14 @@ class BaseSpectralModel():
             self._gauss_std_limits = None
 
 
-    def _reset_data_results(self, clear_freqs=False, clear_spectrum=False, clear_results=False):
-        """Set, or reset, data & results attributes to empty.
+    def _reset_results(self, clear_results=False):
+        """Set, or rest, results attributes to empty.
 
         Parameters
         ----------
-        clear_freqs : bool, optional, default: False
-            Whether to clear frequency attributes.
-        clear_spectrum : bool, optional, default: False
-            Whether to clear power spectrum attribute.
         clear_results : bool, optional, default: False
             Whether to clear model results attributes.
         """
-
-        if clear_freqs:
-            self.freqs = None
-            self.freq_range = None
-            self.freq_res = None
-
-        if clear_spectrum:
-            self.power_spectrum = None
 
         if clear_results:
 
@@ -327,104 +288,21 @@ class BaseSpectralModel():
             self._peak_fit = None
 
 
-    def _prepare_data(self, freqs, power_spectrum, freq_range, spectra_dim=1):
-        """Prepare input data for adding to current object.
+    def _reset_data_results(self, clear_freqs=False, clear_spectrum=False, clear_results=False):
+        """Set, or reset, data & results attributes to empty.
 
         Parameters
         ----------
-        freqs : 1d array
-            Frequency values for the power_spectrum, in linear space.
-        power_spectrum : 1d or 2d array
-            Power values, which must be input in linear space.
-            1d vector, or 2d as [n_power_spectra, n_freqs].
-        freq_range : list of [float, float]
-            Frequency range to restrict power spectrum to.
-            If None, keeps the entire range.
-        spectra_dim : int, optional, default: 1
-            Dimensionality that the power spectra should have.
-
-        Returns
-        -------
-        freqs : 1d array
-            Frequency values for the power_spectrum, in linear space.
-        power_spectrum : 1d or 2d array
-            Power spectrum values, in log10 scale.
-            1d vector, or 2d as [n_power_specta, n_freqs].
-        freq_range : list of [float, float]
-            Minimum and maximum values of the frequency vector.
-        freq_res : float
-            Frequency resolution of the power spectrum.
-
-        Raises
-        ------
-        DataError
-            If there is an issue with the data.
-        InconsistentDataError
-            If the input data are inconsistent size.
+        clear_freqs : bool, optional, default: False
+            Whether to clear frequency attributes.
+        clear_spectrum : bool, optional, default: False
+            Whether to clear power spectrum attribute.
+        clear_results : bool, optional, default: False
+            Whether to clear model results attributes.
         """
 
-        # Check that data are the right types
-        if not isinstance(freqs, np.ndarray) or not isinstance(power_spectrum, np.ndarray):
-            raise DataError("Input data must be numpy arrays.")
-
-        # Check that data have the right dimensionality
-        if freqs.ndim != 1 or (power_spectrum.ndim != spectra_dim):
-            raise DataError("Inputs are not the right dimensions.")
-
-        # Check that data sizes are compatible
-        if freqs.shape[-1] != power_spectrum.shape[-1]:
-            raise InconsistentDataError("The input frequencies and power spectra "
-                                        "are not consistent size.")
-
-        # Check if power values are complex
-        if np.iscomplexobj(power_spectrum):
-            raise DataError("Input power spectra are complex values. "
-                            "Model fitting does not currently support complex inputs.")
-
-        # Force data to be dtype of float64
-        #   If they end up as float32, or less, scipy curve_fit fails (sometimes implicitly)
-        if freqs.dtype != 'float64':
-            freqs = freqs.astype('float64')
-        if power_spectrum.dtype != 'float64':
-            power_spectrum = power_spectrum.astype('float64')
-
-        # Check frequency range, trim the power_spectrum range if requested
-        if freq_range:
-            freqs, power_spectrum = trim_spectrum(freqs, power_spectrum, freq_range)
-
-        # Check if freqs start at 0 and move up one value if so
-        #   Aperiodic fit gets an inf if freq of 0 is included, which leads to an error
-        if freqs[0] == 0.0:
-            freqs, power_spectrum = trim_spectrum(freqs, power_spectrum, [freqs[1], freqs.max()])
-            if self.verbose:
-                print("\nFITTING WARNING: Skipping frequency == 0, "
-                      "as this causes a problem with fitting.")
-
-        # Calculate frequency resolution, and actual frequency range of the data
-        freq_range = [freqs.min(), freqs.max()]
-        freq_res = freqs[1] - freqs[0]
-
-        # Log power values
-        power_spectrum = np.log10(power_spectrum)
-
-        ## Data checks - run checks on inputs based on check modes
-
-        if self._check_freqs:
-            # Check if the frequency data is unevenly spaced, and raise an error if so
-            freq_diffs = np.diff(freqs)
-            if not np.all(np.isclose(freq_diffs, freq_res)):
-                raise DataError("The input frequency values are not evenly spaced. "
-                                "The model expects equidistant frequency values in linear space.")
-        if self._check_data:
-            # Check if there are any infs / nans, and raise an error if so
-            if np.any(np.isinf(power_spectrum)) or np.any(np.isnan(power_spectrum)):
-                error_msg = ("The input power spectra data, after logging, contains NaNs or Infs. "
-                             "This will cause the fitting to fail. "
-                             "One reason this can happen is if inputs are already logged. "
-                             "Input data should be in linear spacing, not log.")
-                raise DataError(error_msg)
-
-        return freqs, power_spectrum, freq_range, freq_res
+        super()._reset_data(clear_freqs, clear_spectrum)
+        self._reset_results(clear_results)
 
 
     def _check_width_limits(self):
